More CI
This change brings the CI for this project to closer alignment with that
of `flightctl`.  It adds linting and the ability to run the linter
locally during development.

One of the new CI tests checks for import order, so a number of the
source files have been updated so that that test now passes.