Fix CI UT issue
In the CI environment, there is no global `user.name` and `user.email`
configured.  As a result, a unit test trying to do a `commit` will fail
if it doesn't set the `user.name` and `user.email`.

This change updates `TestSwitchToBranch_NonExistentBranch` to explicitly
set those values in project-local scope.