Extend logging in services/github.go
This change adds additional logging of the output of the various `git`
commands when the log level is `Debug`.  This will enable us to examine
the log and better understand what happened when something goes wrong.

Also, lint started complaining about some new things. They're related
to it not understanding that `t.Fatal()` terminates test execution, so
I added suppressions.